chore(Dropdown): Refactor unit tests and import barrel
