+++ components/dropdown/index.tsx
@@ -1,26 +1,36 @@
 export { Dropdown, DropdownMenu } from './Dropdown';
 export type { DropdownMenuProps, DropdownProps } from './Dropdown';
-export {
-  DropdownItemAction,
-  DropdownItemCustom,
-  DropdownItemDivider,
-  DropdownItemExpandable,
-  DropdownItemGroup,
-  DropdownItemHeader,
-  DropdownItemList,
-  DropdownItemMultiselect,
-  DropdownItemSelect,
-} from './DropdownItem';
+
+export { DropdownItemAction } from './DropdownItemAction';
 export type {
+  DropdownItemActionButtonProps,
+  DropdownItemActionLinkProps,
   DropdownItemActionProps,
-  DropdownItemCustomProps,
-  DropdownItemDividerProps,
-  DropdownItemExpandableProps,
-  DropdownItemGroupProps,
-  DropdownItemHeaderProps,
-  DropdownItemListProps,
-  DropdownItemMultiselectProps,
-  DropdownItemSelectProps,
-} from './DropdownItem';
+} from './DropdownItemAction';
+
+export { DropdownItemCustom } from './DropdownItemCustom';
+export type { DropdownItemCustomProps } from './DropdownItemCustom';
+
+export { DropdownItemDivider } from './DropdownItemDivider';
+export type { DropdownItemDividerProps } from './DropdownItemDivider';
+
+export { DropdownItemExpandable } from './DropdownItemExpandable';
+export type { DropdownItemExpandableProps } from './DropdownItemExpandable';
+
+export { DropdownItemGroup } from './DropdownItemGroup';
+export type { DropdownItemGroupProps } from './DropdownItemGroup';
+
+export { DropdownItemHeader } from './DropdownItemHeader';
+export type { DropdownItemHeaderProps } from './DropdownItemHeader';
+
+export { DropdownItemList } from './DropdownItemList';
+export type { DropdownItemListProps } from './DropdownItemList';
+
+export { DropdownItemMultiselect } from './DropdownItemMultiselect';
+export type { DropdownItemMultiselectProps } from './DropdownItemMultiselect';
+
+export { DropdownItemSelect } from './DropdownItemSelect';
+export type { DropdownItemSelectProps } from './DropdownItemSelect';
+
 export { DropdownTrigger } from './DropdownTrigger';
 export type { DropdownTriggerProps } from './DropdownTrigger';
